Adapt scripts to new grabber

+++ src/static/static/home/yi-hack/script/wd.sh
@@ -40,18 +40,23 @@ check_rtsp()
         #  echo "$(date +'%Y-%m-%d %H:%M:%S') - Checking RTSP process..." >> $LOG_FILE
         LISTEN=`netstat -an 2>&1 | grep ":$RTSP_PORT_NUMBER " | grep LISTEN | grep -c ^`
         SOCKET=`netstat -an 2>&1 | grep ":$RTSP_PORT_NUMBER " | grep ESTABLISHED | grep -c ^`
-        CPU_1=`top -b -n 2 -d 1 | grep rRTSPServer | grep -v grep | tail -n 1 | awk '{print $8}'`
+        CPU_1=`top -b -n 2 -d 1 | grep h264grabber2 | grep -v grep | tail -n 1 | awk '{print $8}'`
+        CPU_2=`top -b -n 2 -d 1 | grep rRTSPServer | grep -v grep | tail -n 1 | awk '{print $8}'`
 
         if [ $LISTEN -eq 0 ]; then
             echo "$(date +'%Y-%m-%d %H:%M:%S') - Restarting rtsp process" >> $LOG_FILE
             killall -q rRTSPServer
+            killall h264grabber
+            killall h264grabber2
             sleep 1
             restart_rtsp
         fi
         if [[ $(get_config RTSP_STREAM) == "low" ]] || [[ $(get_config RTSP_STREAM) == "both" ]]; then
-            if [ "$CPU_1" == "" ]; then
+            if [ "$CPU_1" == "" ] || [ "$CPU_2" == "" ]; then
                 echo "$(date +'%Y-%m-%d %H:%M:%S') - No running processes for low res, restarting..." >> $LOG_FILE
                 killall -q rRTSPServer
+                killall h264grabber
+                killall h264grabber2
                 sleep 1
                 restart_rtsp
             fi
@@ -61,18 +66,22 @@ check_rtsp()
             if [ "$CPU_1" == "" ]; then
                 echo "$(date +'%Y-%m-%d %H:%M:%S') - No running processes for high res, restarting..." >> $LOG_FILE
                 killall -q rRTSPServer
+                killall h264grabber
+                killall h264grabber2
                 sleep 1
                 restart_rtsp
             fi
             COUNTER_H=0
         fi
         if [ $SOCKET -gt 0 ]; then
-            if [ "$CPU_1" == "0.0" ]; then
+            if [ "$CPU_1" == "0.0" ] && [ "$CPU_2" == "0.0" ]; then
                 COUNTER_L=$((COUNTER_L+1))
                 echo "$(date +'%Y-%m-%d %H:%M:%S') - Detected possible locked process for low res ($COUNTER_L)" >> $LOG_FILE
                 if [ $COUNTER_L -ge $COUNTER_LIMIT ]; then
                     echo "$(date +'%Y-%m-%d %H:%M:%S') - Restarting processes" >> $LOG_FILE
                     killall -q rRTSPServer
+                    killall h264grabber
+                    killall h264grabber2
                     sleep 1
                     restart_rtsp
                     COUNTER_L=0
@@ -81,12 +90,14 @@ check_rtsp()
                 COUNTER_L=0
             fi
 
-            if [ "$CPU_1" == "0.0" ]; then
+            if [ "$CPU_1" == "0.0" ] && [ "$CPU_2" == "0.0" ]; then
                 COUNTER_H=$((COUNTER_H+1))
                 echo "$(date +'%Y-%m-%d %H:%M:%S') - Detected possible locked process for high res ($COUNTER_H)" >> $LOG_FILE
                 if [ $COUNTER_H -ge $COUNTER_LIMIT ]; then
                     echo "$(date +'%Y-%m-%d %H:%M:%S') - Restarting processes" >> $LOG_FILE
                     killall -q rRTSPServer
+                    killall h264grabber
+                    killall h264grabber2
                     sleep 1
                     restart_rtsp
                     COUNTER_H=0
